refactor: moved rtk get from home to cardlist component
